{internal=>tools}/bezier: move package
